dashboard_proxy/requirements.txt
fastapi==0.124.0
uvicorn==0.38.0
google-cloud-storage==3.6.0
